More docs about the caching system
Essentially changes in the underlying data to a MergedOptions won't be
seen by the MergedOptions and all changes should be made to the
MergedOptions instead of the underlying data.

The one "exception" to this is that because when you access the key on a
MergedOptions, you are returned a bookmark of sorts into the
MergedOptions, calling as_dict on that bookmark will return you changes
that have been made in the underlying data.

Note that there are exceptions to that exception and it's best just not
to expect changes in the underlying data to make a difference on the
values given back from the MergedOptions