VersionedDict doesn't actually give us anything
It's never given to the user of MergedOptions because it's a prefixed
class (i.e. accessing one gives us back a MergedOptions "bookmark".

Because of this, it never gets edited directly and it's version never
changes.

Also, the places where I was instantiating them manually meant that the
versioned decorators weren't seeing those objects and so their changing
version number was unecessary.